nexus/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-game-automation"
version = "0.1.0"
description = "Modern Game Automation Framework with AI Integration"
authors = [
    { name = "Nexus Team", email = "nexus@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "pyyaml>=6.0",
    "toml>=0.10.2",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "structlog>=23.2.0",
    "gymnasium>=0.29.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",
    "easyocr>=1.7.0",
    "aiofiles>=23.2.0",
    "httpx>=0.25.0",
    "watchdog>=3.0.0",
    "redis>=5.0.0",
    "psutil>=5.9.0",
    "click>=8.1.0",
    "rich>=13.6.0",
    "pywin32>=306; sys_platform == 'win32'",
    "pydantic>=2.0.0",
    "asyncio>=3.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

ml = [
    "stable-baselines3>=2.1.0",
    "tensorboard>=2.14.0",
    "optuna>=3.4.0",
    "transformers>=4.35.0",
]

[project.scripts]
nexus = "nexus.cli:main"

[tool.setuptools]
packages = ["nexus"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "A", "C4", "T10", "T20"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true